pyproject.toml
[project]
name = "label-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "pyinstaller>=6.14.2",
]
